sub kategori pengepul done - next to laporan pengepul and fixing dashboard

+++ project-ecoranger/Models/Transaksi.cs
@@ -15,7 +15,7 @@ namespace project_ecoranger.Models
         public decimal beratSampah { get; set; }
         public decimal hargaSampah { get; set; }
         public string  StatusTransaksi { get; set; }
-        public int totalTransaksi { get; set; }
+        public decimal totalTransaksi { get; set; }
 
     }
 }
